refactor(deploy-check): banner sections become real modules
Replace the five \`// ---\` comment-banner sections in main.rs with
sibling modules of the same names: \`contract\`, \`hook\`,
\`version_literal\`, \`manifest\`, and \`git\`. main.rs is now just
the CLI definition, the \`run\` dispatcher, and the \`mod\` /
\`use\` lines.

Visibility: items needed across modules are \`pub(crate)\`; helper
types and functions that are local to one module stay private.
\`PACKAGE_NAME\` moves to a \`pub(crate) const\` at the crate root,
referenced by \`manifest\` (lockfile filter) and \`error\` (display
strings) via \`use super::PACKAGE_NAME\`.

Drop the parser-combinator-convention pointer from the
\`version_literal\` module doc — the take_*-style is self-evident
from the code, and the file no longer needs to advertise where
the convention is defined.

Add a test (\`version_only_diff_accepts_arbitrary_distinct_versions\`)
that exercises four non-monotonic pairs (rc.36→rc.20, 0.1.0→0.0.0-rc.88,
0.0.1→0.0.0-rc.72, 0.1.2→0.2.3). The contract requires the two
versions to differ, not to be ordered, and the comment on
\`verify_version_bump_file\` now says so explicitly.
